Progress on fixing the animation bug
